Remove UIParent, add IsSystemWebViewAvailable to PCA
