+++ src/Microsoft.Identity.Client/Platforms/Mac/MacPlatformProxy.cs
@@ -135,6 +135,8 @@ namespace Microsoft.Identity.Client.Platforms.Mac
            () => NetworkInterface.GetAllNetworkInterfaces().Where(nic => nic.OperationalStatus == OperationalStatus.Up)
                                  .Select(nic => nic.GetPhysicalAddress()?.ToString()).FirstOrDefault());
 
+        public override bool IsSystemWebViewAvailable => false;
+
         /// <summary>
         /// Considered PII. Please ensure that it is hashed.
         /// </summary>
