+++ src/Microsoft.Identity.Client/Platforms/net45/NetDesktopPlatformProxy.cs
@@ -73,6 +73,8 @@ namespace Microsoft.Identity.Client.Platforms.net45
             }
         }
 
+        public override bool IsSystemWebViewAvailable => false;
+
         /// <summary>
         ///     Get the user logged in to Windows or throws
         /// </summary>
