+++ src/Microsoft.Identity.Client/Platforms/uap/UapPlatformProxy.cs
@@ -38,12 +38,9 @@ using Windows.System;
 using Microsoft.Identity.Client.Cache;
 using Microsoft.Identity.Client.Core;
 using Microsoft.Identity.Client.Exceptions;
-using Microsoft.Identity.Client.Internal;
 using Microsoft.Identity.Client.UI;
-using String = System.String;
 using Microsoft.Identity.Client.PlatformsCommon.Interfaces;
 using Microsoft.Identity.Client.PlatformsCommon.Shared;
-using Windows.Foundation.Collections;
 using Microsoft.Identity.Client.Mats.Internal;
 
 namespace Microsoft.Identity.Client.Platforms.uap
@@ -57,6 +54,8 @@ namespace Microsoft.Identity.Client.Platforms.uap
             : base(logger)
         {
         }
+
+        public override bool IsSystemWebViewAvailable => false;
 
         /// <summary>
         /// Get the user logged in to Windows or throws
