+++ src/Microsoft.Identity.Client/PlatformsCommon/Shared/AbstractPlatformProxy.cs
@@ -60,6 +60,8 @@ namespace Microsoft.Identity.Client.PlatformsCommon.Shared
             _platformLogger = new Lazy<IPlatformLogger>(InternalGetPlatformLogger);
         }
 
+        public abstract bool IsSystemWebViewAvailable { get; }
+
         protected IWebUIFactory OverloadWebUiFactory { get; set; }
         protected IFeatureFlags OverloadFeatureFlags { get; set; }
